requirements.txt
streamlit==1.45.0
Pillow==11.2.1
torch==2.7.1+cu118
torchvision==0.22.1+cu118
